+++ app/lib/config.js
@@ -1,6 +1,7 @@
 const version = require('../../package').version;
 
-const outputFile = 'pharmacy-data';
+const outputFile = process.env.OUTPUT_FILE || 'pharmacy-data';
+const seedIdFile = process.env.SEED_ID_FILE || 'pharmacy-seed-ids';
 
 module.exports = {
   cacheDataFilename: 'cache-data.json',
@@ -14,7 +15,7 @@ module.exports = {
   outputFilename: `${outputFile}.json`,
   processedPagesFilename: 'processed-pages.json',
   saveEvery: 100,
-  seedIdFile: 'pharmacy-seed-ids',
+  seedIdFile,
   summaryFilename: 'summary.json',
   syndicationApiUrl: 'https://v1.syndication.nhschoices.nhs.uk/organisations/pharmacies',
   version,